+++ corehq/apps/users/tasks.py
@@ -1,10 +1,4 @@
-import functools
 from datetime import datetime
-
-from django.urls import reverse
-from django.utils.html import format_html
-from django.utils.safestring import mark_safe
-from django.utils.translation import ugettext as _
 
 from celery.exceptions import MaxRetriesExceededError
 from celery.schedules import crontab
@@ -12,61 +6,39 @@ from celery.task import task
 from celery.task.base import periodic_task
 from celery.utils.log import get_task_logger
 from couchdbkit import BulkSaveError, ResourceConflict
+from django.urls import reverse
+from django.utils.html import format_html
+from django.utils.safestring import mark_safe
+from django.utils.translation import ugettext as _
 
 from casexml.apps.case.mock import CaseBlock
 from casexml.apps.case.xform import get_case_ids_from_form
-from couchforms.exceptions import UnexpectedDeletedXForm
-from dimagi.utils.couch.bulk import BulkFetchException
-from dimagi.utils.logging import notify_exception
-from soil import DownloadBase
-
 from corehq import toggles
 from corehq.apps.domain.models import Domain
+from corehq.apps.user_importer.tasks import import_users_and_groups
 from corehq.form_processor.exceptions import CaseNotFound, NotAllowed
 from corehq.form_processor.interfaces.dbaccessors import (
     CaseAccessors,
     FormAccessors,
 )
 from corehq.form_processor.models import UserArchivedRebuild
+from couchforms.exceptions import UnexpectedDeletedXForm
+from dimagi.utils.couch.bulk import BulkFetchException
+from dimagi.utils.logging import notify_exception
+from soil import DownloadBase
 
 logger = get_task_logger(__name__)
 
 
 @task(serializer='pickle')
 def bulk_upload_async(domain, user_specs, group_specs):
-    from corehq.apps.users.bulkupload import create_or_update_users_and_groups, create_or_update_groups
-    task = bulk_upload_async
-    DownloadBase.set_progress(task, 0, 100)
-
-    total = len(user_specs) + len(group_specs)
-    DownloadBase.set_progress(task, 0, total)
-
-    group_memoizer, group_results = create_or_update_groups(domain, group_specs)
-
-    DownloadBase.set_progress(task, len(group_specs), total)
-
-    def _update_progress(value, start=0):
-        DownloadBase.set_progress(task, start + value, total)
-
-    user_results = create_or_update_users_and_groups(
-        domain,
-        user_specs,
-        group_memoizer=group_memoizer,
-        update_progress=functools.partial(_update_progress, start=len(group_specs))
-    )
-    results = {
-        'errors': group_results['errors'] + user_results['errors'],
-        'rows': user_results['rows']
-    }
-    DownloadBase.set_progress(task, total, total)
-    return {
-        'messages': results
-    }
+    # remove this after deploying `import_users_and_groups`
+    return import_users_and_groups(domain, user_specs, group_specs)
 
 
 @task(serializer='pickle')
 def bulk_download_users_async(domain, download_id, user_filters):
-    from corehq.apps.users.bulkupload import dump_users_and_groups, GroupNameError
+    from corehq.apps.users.bulk_download import dump_users_and_groups, GroupNameError
     errors = []
     try:
         dump_users_and_groups(
